Split MainActivityTest into base class
