+++ app/src/androidTest/java/com/alloydflanagan/hexcalcrpn/main/MainActivityInfTest.kt
@@ -1,0 +1,203 @@
+package com.alloydflanagan.hexcalcrpn.main
+
+
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.IdRes
+import androidx.test.InstrumentationRegistry
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.ViewInteraction
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.*
+import androidx.test.filters.LargeTest
+import androidx.test.rule.ActivityTestRule
+import androidx.test.runner.AndroidJUnit4
+import com.alloydflanagan.hexcalcrpn.R
+import kotlinx.android.synthetic.main.activity_main.*
+import org.hamcrest.Description
+import org.hamcrest.Matcher
+import org.hamcrest.Matchers.allOf
+import org.hamcrest.TypeSafeMatcher
+import org.junit.Assert
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@LargeTest
+@RunWith(AndroidJUnit4::class)
+class MainActivityInfTest: MainActivityTest() {
+
+    @Rule
+    @JvmField
+    var mActivityTestRule = ActivityTestRule(MainActivity::class.java)
+
+    /**
+     * clears both current value and stack outputs. Prevents failed test from affecting others.
+     */
+    @Before
+    fun clearOutputs() {
+        /** guard against infinite loop if other checks fail */
+        var limit = 15
+        val act = mActivityTestRule.activity
+        while (act.tv_current.text != "0" || act.tv_output.text != "") {
+            enterKeys("c")
+            if (--limit < 0) break
+        }
+    }
+
+    /**
+     * Verify we can get application context, and it is what I think it is.
+     */
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        Assert.assertEquals("com.alloydflanagan.hexcalcrpn", appContext.packageName)
+    }
+
+    /**
+     * Verify that pressing a series of digits creates a number in the current value register,
+     * and that pressing "clear" resets that number to 0.
+     */
+    @Test
+    fun testEntry() {
+        enterKeys("8")
+        checkCurrentIs("8")
+        enterKeys("FA6540123456789ABCDEF")
+
+        checkCurrentIs("8FA6540123456789ABCDEF")
+
+        enterKeys("c")
+        checkCurrentIs("0")
+    }
+
+    /**
+     * Verify that we can do simple addition.
+     */
+    @Test
+    fun testAdd() {
+        enterKeys("1234")
+        enter()
+
+        enterKeys("ABC")
+        enter()
+
+        enterKeys("+")
+
+        checkOutputIs("1CF0")
+    }
+
+    /**
+     * Verify that we can do simple subtraction.
+     */
+    @Test
+    fun testSubtract() {
+        enterKeys("1234")
+        enter()
+
+        enterKeys("ABC")
+        enter()
+        enterKeys("-")
+
+        checkOutputIs("778")
+    }
+
+    @Test
+    fun testMultiply() {
+        enterKeys("AC57")
+        enter()
+
+        enterKeys("FACE")
+        enter()
+
+        checkOutputIs("AC57\nFACE")
+
+        enterKeys("*")
+        checkOutputIs("A8D7A402")
+    }
+
+    @Test
+    fun testDivide() {
+        enterKeys("A8D7A402")
+        enter()
+        enterKeys("AC57")
+        enter()
+
+        checkOutputIs("A8D7A402\nAC57")
+
+        enterKeys("/")
+        checkOutputIs("FACE")
+    }
+
+    @Test
+    fun testAND() {
+        enterKeys("FACE")
+        enter()
+        enterKeys("AC57")
+        enter()
+
+        checkOutputIs("FACE\nAC57")
+
+        enterKeys("&")
+        checkOutputIs("A846")
+    }
+
+    @Test
+    fun testOR() {
+        enterKeys("FACE")
+        enter()
+        enterKeys("AC57")
+        enter()
+
+        checkOutputIs("FACE\nAC57")
+
+        enterKeys("|")
+        checkOutputIs("FEDF")
+    }
+
+    /**
+     * Verify that pressing "c" clears the current value if any, otherwise removes one entry from
+     * stack.
+     */
+    @Test
+    fun testClear() {
+        enterKeys("1234")
+        enter()
+        checkCurrentIs("0")
+        checkOutputIs("1234")
+
+        enterKeys("ABC")
+        enter()
+        checkCurrentIs("0")
+        checkOutputIs("1234\nABC")
+
+        enterKeys("747")
+        checkCurrentIs("747")
+
+        enterKeys("c")
+
+        checkCurrentIs("0")
+        checkOutputIs("1234\nABC")
+
+        enterKeys("c")
+
+        checkOutputIs("1234")
+
+        enterKeys("c")
+
+        checkOutputIs("")
+    }
+
+    /**
+     * Verify pressing "clear" with no input and no stack does nothing.
+     */
+    @Test
+    fun testClearNull() {
+        enterKeys("c") // should do nothing
+        checkOutputIs("")
+        checkCurrentIs("0")
+    }
+
+}
